+++ lib/features/book_detail/presentation/components/book_detail_update_section.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 
 import '../../../../core/theme/app_colors.dart';
-import '../../../../core/theme/app_radius.dart';
 import '../../../../core/theme/app_sizes.dart';
 import '../../../../core/theme/app_spacing.dart';
 import '../../../../core/theme/app_text_styles.dart';
@@ -28,10 +27,6 @@ class BookDetailUpdateSection extends StatelessWidget {
     return Container(
       width: double.infinity,
       padding: const EdgeInsets.all(AppSizes.bookDetailUpdateSectionPadding),
-      decoration: BoxDecoration(
-        color: AppColors.bookDetailUpdateSectionBackground,
-        borderRadius: BorderRadius.circular(AppRadius.lg),
-      ),
       child: Column(
         children: [
           for (var i = 0; i < entries.length; i++)